fix: userAtomをuserに変更
